+++ extensions/gif-search/src/tenor.tsx
@@ -1,19 +1,43 @@
-import { useEffect } from "react";
+import { useEffect, useReducer } from "react";
+
 import { List, showToast, Toast } from "@raycast/api";
 
 import { getShowPreview, GIF_SERVICE } from "./preferences";
+
+import AppContext, { initialState, reduceAppState } from "./components/AppContext";
 import { GifList } from "./components/GifList";
+import { FavesList } from "./components/FavesList";
+
+import useFavorites from "./hooks/useFavorites";
 import useSearchAPI from "./hooks/useSearchAPI";
 
 import "./fetch-polyfill";
 
 export default function Command() {
+  const serviceName = GIF_SERVICE.TENOR;
   const showPreview = getShowPreview();
+
   const [results, isLoading, setSearchService, setSearchTerm] = useSearchAPI({});
+  const [favIds, favItems, isFavsLoading, loadFavs, populate] = useFavorites({});
+
+  const [state, dispatch] = useReducer(reduceAppState, initialState);
 
   useEffect(() => {
-    setSearchService(GIF_SERVICE.TENOR);
-  }, []);
+    setSearchService(serviceName);
+    loadFavs(serviceName);
+  }, [loadFavs]);
+
+  useEffect(() => {
+    if (favIds?.ids) {
+      dispatch({ type: "set", ids: favIds?.ids, service: serviceName });
+    }
+  }, [favIds]);
+
+  useEffect(() => {
+    if (state.favIds) {
+      populate(state.favIds, serviceName);
+    }
+  }, [state]);
 
   useEffect(() => {
     if (results?.error) {
@@ -26,14 +50,17 @@ export default function Command() {
   }, [results?.error]);
 
   return (
-    <List
-      isLoading={isLoading}
-      throttle={true}
-      searchBarPlaceholder="Search for gifs on Tenor..."
-      onSearchTextChange={setSearchTerm}
-      isShowingDetail={showPreview && results?.items?.length != 0}
-    >
-      <GifList term={results?.term} results={results?.items} />
-    </List>
+    <AppContext.Provider value={{ state, dispatch }}>
+      <List
+        isLoading={isLoading || isFavsLoading}
+        throttle={true}
+        searchBarPlaceholder="Search for gifs on Tenor..."
+        onSearchTextChange={setSearchTerm}
+        isShowingDetail={showPreview && (results?.items?.length ?? 0) + (favItems?.items?.length ?? 0) != 0}
+      >
+        <FavesList results={favItems?.items} show={!!favItems?.items?.length && !results?.term} />
+        <GifList term={results?.term} results={results?.items} />
+      </List>
+    </AppContext.Provider>
   );
 }